firmware: qcom_scm: Use fixed width src vm bitmap
Bugzilla: https://bugzilla.redhat.com/2213601
Conflicts: fastrpc.c in CS9 is not utilizing qcom_scm, so related code was,
removed. Downstream and Upstream do not match 1:1, so changes are made across
different lines.

commit 968a26a07f75377afbd4f7bb18ef587a1443c244
Author: Elliot Berman <quic_eberman@quicinc.com>
Date:   Mon Feb 13 10:18:29 2023 -0800

    firmware: qcom_scm: Use fixed width src vm bitmap

    The maximum VMID for assign_mem is 63. Use a u64 to represent this
    bitmap instead of architecture-dependent "unsigned int" which varies in
    size on 32-bit and 64-bit platforms.

    Acked-by: Kalle Valo <kvalo@kernel.org> (ath10k)
    Tested-by: Gokul krishna Krishnakumar <quic_gokukris@quicinc.com>
    Link: https://lore.kernel.org/r/20230213181832.3489174-1-quic_eberman@quicinc.com

(cherry picked from commit 14291453f09a50e70c7129e59cef62690d94796d)

+++ drivers/remoteproc/qcom_q6v5_mss.c
@@ -201,8 +201,8 @@ struct q6v5 {
 	bool has_alt_reset;
 	bool has_mba_logs;
 	bool has_spare_reg;
-	int mpss_perm;
-	int mba_perm;
+	u64 mpss_perm;
+	u64 mba_perm;
 	const char *hexagon_mdt_image;
 	int version;
 };
@@ -377,7 +377,7 @@ static void q6v5_pds_disable(struct q6v5 *qproc, struct device **pds,
 	}
 }
 
-static int q6v5_xfer_mem_ownership(struct q6v5 *qproc, int *current_perm,
+static int q6v5_xfer_mem_ownership(struct q6v5 *qproc, u64 *current_perm,
 				   bool local, bool remote, phys_addr_t addr,
 				   size_t size)
 {
@@ -805,7 +805,7 @@ static int q6v5_mpss_init_image(struct q6v5 *qproc, const struct firmware *fw)
 	unsigned long dma_attrs = DMA_ATTR_FORCE_CONTIGUOUS;
 	dma_addr_t phys;
 	void *metadata;
-	int mdata_perm;
+	u64 mdata_perm;
 	int xferop_ret;
 	size_t size;
 	void *ptr;
